modify the average of list

+++ avgoflst.cpp
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-int avgoflst(int lst[], int size);
+double avgoflst(int lst[], int size);
 
 int main() {
 	int lst[] = {19, 2, 20, 1, 0, 18};
@@ -10,7 +10,7 @@ int main() {
 	return 0;
 }
 
-int avgoflst(int lst[], int size) {
+double avgoflst(int lst[], int size) {
 	int sum;
 	for (int i = 0; i < size; i++) {
 		sum += lst[i];
